The window that controls the arduino Light presets
This window can be opened from the Mainwindow through a button. When opened, this script controls the light settings window. This script can save light sequence presets to a file, display saved light presets, and update these when a new song has been detected. It will also send the light settings to the arduino device